+++ LenovoLegionToolkit.WPF/Controls/Dashboard/DashboardGroupControl.cs
@@ -1,5 +1,6 @@
 ﻿using System.Windows;
 using System.Windows.Controls;
+using LenovoLegionToolkit.WPF.Extensions;
 
 namespace LenovoLegionToolkit.WPF.Controls.Dashboard;
