+++ LenovoLegionToolkit.WPF/Controls/Dashboard/Edit/EditDashboardItemControl.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Windows.Controls;
+using LenovoLegionToolkit.WPF.Extensions;
 using Wpf.Ui.Common;
 using Wpf.Ui.Controls;
 using Button = Wpf.Ui.Controls.Button;
