+++ LenovoLegionToolkit.WPF/Enums.cs
@@ -1,0 +1,29 @@
+﻿namespace LenovoLegionToolkit.WPF;
+
+public enum DashboardGroupType
+{
+    Power,
+    Graphics,
+    Display,
+    Other,
+    Custom
+}
+
+public enum DashboardItem
+{
+    PowerMode,
+    BatteryMode,
+    AlwaysOnUsb,
+    HybridMode,
+    DiscreteGpu,
+    Resolution,
+    RefreshRate,
+    DpiScale,
+    Hdr,
+    OverDrive,
+    Microphone,
+    FlipToStart,
+    TouchpadLock,
+    FnLock,
+    WinKeyLock
+}
