+++ LenovoLegionToolkit.WPF/Windows/Dashboard/EditDashboardWindow.xaml.cs
